pyproject.toml
[project]
name = "japan-stock-ai-trader"
version = "0.1.0"
dependencies = [
    "requests",
    "yfinance",
    "pandas",
    "numpy",
    "requests",
    "PyYAML",
]
requires-python = ">=3.13"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]